pyproject.toml
[tool.poetry]
name = "boteco"
version = "0.1.0"
description = ""
authors = ["Marco Rougeth <rougeth@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.10.7"
"discord.py" = "^2.0.1"
loguru = "^0.6.0"
python-slugify = "^6.1.2"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
ipython = "^8.5.0"
black = "^22.8.0"
pytest = "^7.1.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
